Introducing an important new file to help devs
This .env file can be used to configure your local environment in an automated way, without requiring you to put variables in a profile or startup script on your machine. This is super useful for security keys to the BrowserStack testing cloud and more. And it means our scripts can work out of the box without having to hard code any values into them or touch your profile at all.