Complete redesign and enable of functional tests.
Raising the quality bar quite a bit here. Refactoring to reduce client-side injection. Making use of real keyboard input to check form values. Improving performance. Implementing skip() so that tests no longer need to be commented out when they are not production-ready. DRY-ing things up a lot. And preparing for Intern 3 and moving to the Page Object pattern.