Fixing module API reference.
This fixes the expectations set by the module loader, as per the design of the core. All modules should behave this way. The module is not responsible for creating the object that will represent it, that job is strictly in the hands of the module loader, so that it can dictate how and where the module's API gets used in other places, and so that any design upgrades can be shared across the entire system at once. This bug prevented other modules from properly referencing its API as a dependency. r=anardi