Move legacy tests to more intuitive place
Previously, we had sibling test/ and tests/ directories that were very confusing, but only because we desired to keep the legacy machinery running in tandem with development on the new testing framework. Now that only the new framework is able to function since our switch to AMD, the old machinery is permanently broken, making it pointless to maintain the old directory structure.