Emergency build fix.
The gulp-jshint package specifies a very loose peerDependency on jshint. Meanwhile, jshint just pushed a new minor version which is buggy on old Node (at least 0.10.x). Since our CI is using this ancient version of Node, our builds have been failing to lint in CI, but not locally where everyone has a modern setup. The proper fix is to modernize our CI infrastructure, but this works around the problem by putting jshint versioning under our control. It is not a hack, because npm 3 and up will not install peerDependencies for you anyway, since it is a footgun where you are opting into a situation that may have version conflicts (and therefor you ought to install them manually).