Only ever use local voices. Error out if we cannot.
This patch enhances security at the expense of the user experience. Instead of the letting the browser choose a voice intelligently based on content and computer language, online status, etc. we override the voice in order to guarantee that the voice never transmits data over a network. If we cannot find local voices, we now abort speaking completely.