Implement locale negotiation.
This takes care of the edge case where sitecues tells the native speech system to use a specific locale with an accent but that accent is not available. Before it would just error out, now it peforms locale negotiation, falling back to the user's preferred voice of the given language.