Problem: We are not processing stylesheets with this pattern: <link rel="stylesheet" id="sitecues-XXXX" ...>
Why? Because we think they are stylesheets inserted by sitecues.js, which always uses "sitecues-" as a prefix in the ID. We like the readability of "sitecues-" because it helps websites know where those stylesheets came from.

How to fix? Use sitecues-js- as prefix for our stylesheets' ids instead of sitecues- because sitecues.com actually uses sitecues- as a stylesheet id's prefix. However, "sitecues-js-" is a lot less likely.

r=ben