Improve photo detection.
Was not detecting large photos when they were low quality because the algorithm considered 6 pixels to be multi-use. Better if we require at least 5% of the pixels to be the same color before we mark them as multi use.

Old code: if a color was reused 6 times, it was considered a reused colors
New code: if a color is reused for 5% of the image it is considered a reused color

Both: the more reused colors there are, the less likely the image is to be a photo (this did not change)

r=gradetsky