Sort voices based upon available metadata.
Previously, we were relying upon browsers following the spec and giving us default voices at the beginning of the voice list. In practice, only Chrome does this. Firefox and Safari both give the list sorted alphabetically. As a result, on OS X the horrible Agnes voice was being used in those browsers instead of Alex (the current out-of-the-box default). We now gaurantee that the actual default voice is used. Also fix some edge cases by normalizing locale casing and rejecting promise errors properly.