Fix regression in local player edge case.
When sitecues needs to synthesize speech, we have to be very careful about sorting voices. We need to handle the case where the locale does not include an accent, and the user does not have a voice without an accent. Yet we never want to add an accent in any other case. Ideally we would apply this same logic to cloud speech as well. But that's out of scope here.