SC-3655. Panel focus rectangle does not move in secondary panel, in Firefox.
We actually change focus on tab, but it looks like we don't because the focus outline rect doesn't change for the 4 secondary buttons.

Reason: getBoundingClientRect() is broken unless we use transform in Firefox, vs. CSS transform. Fortunately we already have that option because of IE.

It causes JS animations to be used for the buttons, e.g. the rolling About button. However, they still look good and smooth.
This is the easiest fix. The other option would be to change helper.getRect() to take into account CSS transform. I suggest we try this first and have Alex test it thoroughly.

r=seth