Inform gulp about deployment errors.
The .on() function of a stream doesn’t care what the return value of the provided callback is. The stream and .on() have no knowledge of it being a promise, let alone it being a rejected promise.

We must propogate errors back to gulp by wrapping the whole stream in a promise that will be rejected if any error occurs. An alternative to this would be to `throw` inside of the stream’s end handler. Except that isn’t possible because it would need to be chained off the promise, and promises implicitly convert throws to rejected promises. Then we’re back to square one, because again, the stream doesn’t care about promises (rejected or otherwise).