Remove obsolete hack.
This .mailmap file was previously used to keep us under our user limit for FeCru. Since we no longer use that service, this configuration, this file is now irrelevant. A nice side effect is that GitHub will now show the correct names and email addresses for all commits, since we are no longer telling it to mask them under the Former Developer alias.